war_breakdown_extraction/src/requirements.txt
google-cloud-storage==1.32.0
google-cloud-vision==2.0.0
